+++ projects/ngx-stripe-docs/src/app/examples/payment-element.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewChild, inject } from '@angular/core';
+import { ChangeDetectorRef, Component, OnInit, ViewChild, inject } from '@angular/core';
 import { ReactiveFormsModule, UntypedFormBuilder, Validators } from '@angular/forms';
 
 import { StripePaymentElementComponent, injectStripe } from 'ngx-stripe';
@@ -32,6 +32,7 @@ export default class PaymentElementExampleComponent implements OnInit {
 
   private readonly fb = inject(UntypedFormBuilder);
   private readonly plutoService = inject(NgStrPlutoService);
+  private readonly cdr = inject(ChangeDetectorRef);
 
   stripeTest = this.fb.group({
     name: ['Angular v12', [Validators.required]],
@@ -53,6 +54,7 @@ export default class PaymentElementExampleComponent implements OnInit {
       })
       .subscribe((pi) => {
         this.elementsOptions.clientSecret = pi.client_secret;
+        this.cdr.detectChanges();
       });
   }